Parser: Měly by fungovat základní konstrukce

+++ test/test_examples/syntactic_examples/X_correct_fun_return.zig
@@ -1,5 +1,5 @@
 const ifj = @import("ifj24.zig");
 
-pub fn foo(a: ?f64) ?f64 {
+pub fn foo(a: ?f64) ?i32 {
     return a;
 }
